% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizing.R
\name{day_series}
\alias{day_series}
\title{Time Series}
\usage{
day_series(path)
}
\arguments{
\item{path}{The spreadsheet file path.}
}
\value{
Returns a ggplot2 plot
}
\description{
Plots a time series ( day ) of the number of submissions
}
