% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{department_distribute}
\alias{department_distribute}
\title{Distributing Departments}
\usage{
department_distribute(data_vector)
}
\arguments{
\item{data_vector}{The list produced by general_distribute()}
}
\value{
A List of data frames
}
\description{
A helper function that splits Departments frame into many 100-row frame for each academic year for each department
}
\examples{
\dontrun{
library(dplyr)
data_list <- raw_data \%>\%
                 split_malicious() \%>\%
                 general_department_split() \%>\%
                 general_distribute() \%>\%
                 depatment_distribute()
}

}
