% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizing.R
\name{department_ratio}
\alias{department_ratio}
\title{Visualizing department proprtions}
\usage{
department_ratio(path)
}
\arguments{
\item{path}{The spreadsheet file path.}
}
\value{
Returns a ggplot2 plot
}
\description{
This function plots a stacked bar chart that is consisted of stacked bars,
each bar represents the number of submissions of each department to the whole each stack in a bar
represents the propotion of a year in a department
}
