% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{general_department_split}
\alias{general_department_split}
\title{Splitting Mainstream group from first year to third year.}
\usage{
general_department_split(data_Vector)
}
\arguments{
\item{data_vector}{A list of data frames, must be the output of the split_malicious() function.}
}
\value{
A list of data frames.
}
\description{
A helper function that splits the mainstream students from the rest.
}
\examples{
\dontrun{
library(dplyr)
data_list <- split_malicious(raw_data) \%>\% general_department_split()
}
}
