% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{general_distribute}
\alias{general_distribute}
\title{Distributing the Mainstream}
\usage{
general_distribute(data_vector)
}
\arguments{
\item{data_vector}{The list produced by general_department_split()}
}
\value{
A List of data frames
}
\description{
A helper function that splits the mainstream frame into many 100-row frame for each academic year.
}
\examples{
\dontrun{
library(dplyr)
data_list <- raw_data \%>\%
                 split_malicious() \%>\%
                 general_department_split() \%>\%
                 general_distribute()
}

}
