% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizing.R
\name{malicious_ratio}
\alias{malicious_ratio}
\title{Visualizing malicious/non-malicious submissions ratio}
\usage{
malicious_ratio(path)
}
\arguments{
\item{path}{The spreadsheet file path.}
}
\value{
Returns a ggplot2 plot
}
\description{
This function plots a piechart that plots the stated propotion
}
