% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{pipeline}
\alias{pipeline}
\title{Data Cleaning Pipeline}
\usage{
pipeline(path)
}
\arguments{
\item{path}{the Excel file path}
}
\value{
The list of all data frames.
}
\description{
This is the function you are going to use.
}
