% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{read_data}
\alias{read_data}
\title{Reading data}
\usage{
read_data(path)
}
\arguments{
\item{path}{The spreadsheet file path}
}
\value{
A data frame
}
\description{
A simple helper function that reads excel files.
}
