% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{split_malicious}
\alias{split_malicious}
\title{Splitting Malicious Submissions}
\usage{
split_malicious(data)
}
\arguments{
\item{data}{The RAW data you want to split}
}
\value{
Returns a list of data frames where each type of malice is included in a data frame of its own. While non
malicious data is included in the Malicious_Not data frame.
}
\description{
A helper function that detects common malicious submission casses & marks them as so.
}
