% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{write_data}
\alias{write_data}
\title{Saving Data}
\usage{
write_data(data_vector)
}
\arguments{
\item{data_vector}{The list resulted from department_distribute().}
}
\description{
Saving data into spreadsheets.
}
