% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizing.R
\name{year_ratio}
\alias{year_ratio}
\title{Visualizing academic year proprtions}
\usage{
year_ratio(path)
}
\arguments{
\item{path}{The spreadsheet file path.}
}
\value{
Returns a ggplot2 plot
}
\description{
This function plots a piechart that is consisted of sectors,
each sector represents the ratio of submissions of each academic year to the whole.
}
